IS-148 Backend changes for QR-code
QR-size in UI is now configurable.
The meaning of bankid.ui.display-qr-help has changed to display an intermediate view for the QR-code (including help texts and a button to display the QR-code).
